Move key registrations to a hook
CORL-1334
